GetScore {Tetris}	R Documentation

Win points

Description

Check if the player will win points

Usage


GetScore(tables)

Arguments

tables
a matrix which contains the information where tetrominos locates

Details

This function will check if any lines are filled by same color and player will win 10 points

Author(s)

Tingting & Chang

[Package Tetris version 0.0.0.9000 Index]