MoveDown {Tetris}	R Documentation

Move right

Description

Move down the dropping tetromino.

Usage


MoveDown(cubes, tables)

Arguments

cubes
a matrix which contains the information where tetromino locates

tables
a matrix which contains the information of the current backgroup including previous tetrominos location

Details

This function will move down the dropping tetromino.

Author(s)

Tingting & Chang

[Package Tetris version 0.0.0.9000 Index]