MoveLeft {Tetris}	R Documentation

Move left

Description

Move the dropping tetromino to the left

Usage


MoveLeft(cubes, tables)

Arguments

cubes
a matrix which contains the information where tetromino locates

tables
a matrix which contains the information of the current backgroup including previous tetrominos location

Details

This function will move the dropping tetromino to the left

Author(s)

Tingting & Chang

[Package Tetris version 0.0.0.9000 Index]