checkNext3Block_y {Tetris}	R Documentation

Check tetrominos moving towards

Description

Check if the tetromino should move down 3

Usage


checkNext3Block_y(cubes, tables, direct)

Arguments

cubes
a matrix which contains the information where tetromino locates

tables
a matrix which contains the information of the current backgroup including previous tetrominos location

Details

This function will check if the tetromino can move down 3

Author(s)

Tingting & Chang

[Package Tetris version 0.0.0.9000 Index]