endGame {Tetris}	R Documentation

End Game

Description

Check if the game will ends

Usage


endGame(tables)

Arguments

tables
a matrix which contains the information where tetrominos locates

Details

This function will check if any coloums is filled by same color and the game will be end

Author(s)

Tingting

[Package Tetris version 0.0.0.9000 Index]